Back-end/ai_service/requirements.txt
annotated-types==0.7.0
anyio
certifi==2026.1.4
cryptography
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
google-genai==0.6.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
pydantic
pydantic_core
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
SQLAlchemy==2.0.31
starlette
tqdm==4.66.4
typing_extensions==4.12.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==14.0